Hugo vars now parsing in Sass, fixed author
Huey can now read Site.Param options from the config.toml file.
This allows for users to specify new values without needing to
modify the Sass configuration directly.

THe author footer was hardset as the theme author, which was a bug.
Fixed so now uses the .Param.author in the footer.